+++ src/components/Button.js
@@ -1,0 +1,7 @@
+import React from 'react';
+
+const Button = ({ handleClick, text }) => (
+  <button onClick={handleClick}>{text}</button>
+);
+
+export default Button;
