+++ src/components/CountSeconds.js
@@ -1,0 +1,7 @@
+import React from 'react';
+
+const CountSeconds = ({ seconds }) => (
+  <div>You didn't refresh this page {seconds} seconds :)</div>
+);
+
+export default CountSeconds;
