+++ src/components/Counter.js
@@ -1,0 +1,7 @@
+import React from 'react';
+
+const Counter = ({ counter }) => (
+  <div>Result of clicking is {counter} now :)</div>
+);
+
+export default Counter;
