+++ src/components/Hello.js
@@ -1,0 +1,18 @@
+import React from 'react';
+
+const Hello = ({ name, age }) => {
+  name = 'Lora';
+  age = 26;
+  const bornYear = () => new Date().getFullYear() - age;
+  return (
+    <div>
+      <p>
+        Hello world, my name is {name}, I am {age} years old.
+      </p>
+      <p>So I was born in {bornYear()}.</p>
+      <p>This is my website :)</p>
+    </div>
+  );
+};
+
+export default Hello;
