+++ src/components/Note.js
@@ -1,0 +1,7 @@
+import React from 'react';
+
+const Note = ({ note }) => {
+  return <li>{note.content}</li>;
+};
+
+export default Note;
